refactor: move custom Vite plugins (manifest, robots, sitemap) to separate folder
