+++ vite-plugins/index.ts
@@ -1,0 +1,3 @@
+export * from './manifest';
+export * from './robots';
+export * from './sitemap';
